Unified tests for angular material and ionic
* added new package `angular-test`
* fixed issues found by tests
