+++ packages/angular-material/package.json
@@ -63,6 +63,7 @@
   },
   "devDependencies": {
     "@jsonforms/examples": "^2.0.12-rc.1",
+    "@jsonforms/angular-test": "^2.0.12-rc.1",
     "@types/node": "^10.10.0",
     "angular2-template-loader": "^0.6.2",
     "copy-webpack-plugin": "^4.5.1",
