+++ packages/angular-material/test/boolean-control.spec.ts
@@ -24,19 +24,16 @@
 */
 import { NgRedux } from '@angular-redux/store';
 import { MockNgRedux } from '@angular-redux/store/testing';
-import { DebugElement } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatCheckbox, MatCheckboxModule, MatError, MatFormFieldModule } from '@angular/material';
-import { By } from '@angular/platform-browser';
+import { booleanTest } from '@jsonforms/angular-test';
 import { NOT_APPLICABLE } from '@jsonforms/core';
 import { BooleanControlRenderer, booleanControlTester } from '../src';
 
-const uischema = {
-    type: 'Control',
-    scope: '#/properties/foo'
-};
-
 describe('Material boolean field tester', () => {
+    const uischema = {
+        type: 'Control',
+        scope: '#/properties/foo'
+    };
     it('should fail', () => {
         expect(booleanControlTester(undefined, undefined)).toBe(NOT_APPLICABLE);
         expect(booleanControlTester(null, undefined)).toBe(NOT_APPLICABLE);
@@ -82,279 +79,11 @@ describe('Material boolean field tester', () => {
         ).toBe(2);
     });
 });
-describe('Boolean control', () => {
-    let fixture: ComponentFixture<any>;
-    let checkboxDebugElement: DebugElement;
-    let checkboxNativeElement: HTMLElement;
-    let checkboxInstance: MatCheckbox;
-    let component: BooleanControlRenderer;
-    let labelElement: HTMLLabelElement;
-
-    const data = { foo: true };
-    const schema = {
-        type: 'object',
-        properties: {
-            foo: {
-                type: 'boolean'
-            }
-        }
-    };
-
-    const getImports = () => [
-        MatCheckboxModule,
-        MatFormFieldModule
-    ];
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [BooleanControlRenderer],
-            imports: getImports(),
-            providers: [
-                { provide: NgRedux, useFactory: MockNgRedux.getInstance }
-            ]
-        }).compileComponents();
-
-        MockNgRedux.reset();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(BooleanControlRenderer);
-        component = fixture.componentInstance;
-        checkboxDebugElement = fixture.debugElement.query(By.directive(MatCheckbox));
-        checkboxInstance = checkboxDebugElement.componentInstance;
-        checkboxNativeElement = checkboxDebugElement.nativeElement;
-        labelElement = checkboxNativeElement.querySelector('label');
-    });
-
-    it('should render', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-        expect(checkboxInstance.disabled).toBe(false);
-    });
-
-    it('should update via input event', () => {
-
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        component.ngOnInit();
-
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-
-        const spy = spyOn(component, 'onChange');
-        labelElement.click();
-        // trigger change detection
-        fixture.detectChanges();
-
-        expect(spy).toHaveBeenCalled();
-        expect(checkboxInstance.checked).toBe(false);
-    });
-
-    it('should support updating the state', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data: { foo: false },
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        expect(component.data).toBe(false);
-        expect(checkboxInstance.checked).toBe(false);
-    });
-    it('should update with undefined value', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data: { foo: undefined },
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        expect(component.data).toBe(undefined);
-        expect(checkboxInstance.checked).toBe(false);
-    });
-    it('should update with null value', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data: { foo: null },
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        expect(component.data).toBe(null);
-        expect(checkboxInstance.checked).toBe(false);
-    });
-    it('should not update with wrong ref', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data: { foo: true, bar: false },
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        expect(component.data).toBe(true);
-        expect(checkboxInstance.checked).toBe(true);
-    });
-    it('can be disabled', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-        component.disabled = true;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(checkboxInstance.disabled).toBe(true);
-
-    });
-    it('id should be present in output', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-        component.id = 'myId';
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                }
-            }
-        });
-        mockSubStore.complete();
-        fixture.detectChanges();
-        component.ngOnInit();
-        expect(checkboxInstance.id).toBe('myId');
-
-    });
-
-    it('should display errors', () => {
-        const mockSubStore = MockNgRedux.getSelectorStub();
-        component.uischema = uischema;
-
-        mockSubStore.next({
-            jsonforms: {
-                core: {
-                    data,
-                    schema,
-                    errors: [{
-                        dataPath: 'foo',
-                        message: 'Hi, this is me, test error!'
-                    }]
-                }
-            },
-        });
-        mockSubStore.complete();
-        component.ngOnInit();
-        fixture.detectChanges();
-        const debugErrors: DebugElement[] = fixture.debugElement.queryAll(By.directive(MatError));
-        expect(debugErrors.length).toBe(1);
-        expect(debugErrors[0].nativeElement.textContent).toBe('Hi, this is me, test error!');
-    });
-});
+const imports = [MatCheckboxModule, MatFormFieldModule];
+const providers = [
+    { provide: NgRedux, useFactory: MockNgRedux.getInstance }
+];
+const componentUT: any = BooleanControlRenderer;
+const errorTest = {errorInstance: MatError, numberOfElements: 1, indexOfElement: 0};
+describe('Boolean control',
+         booleanTest(imports, providers, componentUT , MatCheckbox, errorTest, 'label'));
