+++ packages/angular-test/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@jsonforms/angular-test",
+  "version": "2.0.12-rc.1",
+  "private": true,
+  "main": "./lib/index.js",
+  "dependencies": {
+    "@jsonforms/core": "^2.0.12-rc.1",
+    "@jsonforms/angular": "^2.0.12-rc.1",
+    "redux": "^3.7.2"
+  },
+  "devDependencies": {
+    "jasmine": "^3.2.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^3.0.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^3.0.5",
+    "karma-coverage-istanbul-reporter": "^2.0.4"
+  },
+  "scripts": {
+    "build": "../../node_modules/.bin/tsc",
+    "test": "echo 'Nothing to do'"
+  }
+}
